Taranis supports MULTI DSM telemetry Add second XY-curve and a 6-point curve Add individual cells to telemetry items SKY/AR9X support ERC lcd Taranis sync PXX with XJT to reduce latency Taranis allow 9 custom telemetry values Add PXX option for only 8 channels DSM telemetry, add "Use 'A' as RSSI value" Remove ASSAN protocol Improve custom telemetry editing menu Language updates

+++ radio/ersky9x/src/frsky.h
@@ -374,7 +374,7 @@ extern FrskyData frskyTelemetry[4];
 extern int16_t FrskyHubData[] ;
 //extern int16_t FrskyHubMin[] ;
 //extern int16_t FrskyHubMax[] ;
-extern uint8_t FrskyVolts[];
+extern uint16_t FrskyVolts[];
 extern uint8_t FrskyBattCells[] ;
 extern uint8_t FrskyAlarmCheckFlag ;
 //extern uint8_t MaxGpsSpeed ;
